+++ test/PHPUnit/class/helios/HeliosArchiveControlerTest.php
@@ -16,7 +16,6 @@ class HeliosArchiveControlerTest extends S2lowTestCase {
         parent::setUp();
         $this->heliosArchiveControler = $this->getObjectInstancier()->get("HeliosArchiveControler");
         $this->heliosTransactionsSQL = new HeliosTransactionsSQL($this->getSQLQuery());
-
     }
 
     public function testSetArchiveEnAttenteEnvoiSEA(){
@@ -66,7 +65,7 @@ class HeliosArchiveControlerTest extends S2lowTestCase {
 
 	public function testSend(){
 		$transaction_id = $this->setTransactionEnattente();
-		$this->heliosArchiveControler->setPastellFactory($this->getPastellFactory());
+		$this->heliosArchiveControler->setPastellWrapperFactory($this->getPastellFactory());
 		$this->expectOutputRegex("#Impossible d'envoyer la transaction $transaction_id : Erreur renvoyé par le mock#");
 		$this->heliosArchiveControler->sendAllArchive();
 	}
@@ -79,22 +78,22 @@ class HeliosArchiveControlerTest extends S2lowTestCase {
 
 	/**
 	 * @param int $returnCreateActes
-	 * @return PastellFactory
+	 * @return PastellWrapperFactory
 	 */
 	private function getPastellFactory($returnCreateActes = 0){
-		$pastell = $this->getMockBuilder('Pastell')->disableOriginalConstructor()->getMock();
+		$pastell = $this->getMockBuilder('PastellWrapper')->disableOriginalConstructor()->getMock();
 		$pastell->expects($this->any())->method('createActes')->willReturn($returnCreateActes);
 		$pastell->expects($this->any())->method('getLastError')->willReturn("Erreur renvoyé par le mock");
 
 
-		$pastellFactory = $this->getMockBuilder('PastellFactory')->getMock();
+		$pastellFactory = $this->getMockBuilder('PastellWrapperFactory')->getMock();
 		$pastellFactory->expects($this->any())->method('getNewInstance')->willReturn($pastell);
 		return $pastellFactory;
 	}
 
 	public function testSendTransactionEnErreur(){
 		$this->setTransactionEnattente();
-		$this->heliosArchiveControler->setPastellFactory($this->getPastellFactory());
+		$this->heliosArchiveControler->setPastellWrapperFactory($this->getPastellFactory());
 
 		$date = date("Y-m-d",strtotime("-2 days"));